Change statement for One More Thing
added the following statement :
##One more thing!
To help you get motivated, when attending, you need to put in S$50. We are ***NOT*** profiting from this program. We will return the full sum of S$50 back to your provided that you can finish your 30 Day Challenge! The main reason is because we are only ***accepting 20 students*** and we believe that this commitment of S$50 will deter away students who are not serious and to give this opportunity to someone who does.

Signed-off-by: Bryan Lee <leetucksing@gmail.com>