edit: One More Thing
changed:
You need to put in S$50 for this 30 Day iOS Challenge. We are ***NOT*** profiting from this program. We will return the full sum of S$50 back to you, after you finish your 30 Day Challenge! Our team strongly believe that this monetary deposit would further motivate you to commit and complete the challenge.

Signed-off-by: Bryan Lee <leetucksing@gmail.com>